Fixed migration not being reversable
Migration 0005_auto_20151019_1121 is not reversible because the default_author operation had no reverse code. Since the field the operation sets data for is made in this migration, it should be safe to use noop. The error that was thrown was "django.db.migrations.exceptions.IrreversibleError: Operation <RunPython <function default_author at 0x11292f268>> in blog.0005_auto_20151019_1121 is not reversible"

This made the entire package not able to be uninstalled because it was not able to unapply any migrations. With this small fix, I was able to unapply all migrations.